backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
groq
langchain
langchain-community
chromadb
sentence-transformers
pypdf
requests